// src/app/models/Corrida.ts
export class Corrida {
    id: number = 0;
    idcorrida?: number;
    descricao_corrida: string = '';
    data_corrida: string = '';
    distancia5km: boolean = false;
    distancia10km: boolean = false;
    distancia25km: boolean = false;
    distancia_5km?: boolean;
    distancia_10km?: boolean;
    distancia_25km?: boolean;
}